Add working paper template & move bibtex to Suggests
Since the bibtex package is currently Orphaned at CRAN, it causes a
Warning to have it in Imports, so I moved it to Suggests. It still
causes a Note, but this is fine for now (discussion at rOpenSci implies
it won't be Orphaned for long). For more discussion of this issue, see
CRAN policies at https://cran.r-project.org/web/packages/policies.html